browser compatible objects
