+++ packages/ts-moose-lib/src/dmv2/sdk/view.ts
@@ -1,5 +1,5 @@
 import { dropView } from "../../blocks/helpers";
-import { Sql, toStaticQuery } from "../../consumption-apis/helpers";
+import { Sql, toStaticQuery } from "../../sqlHelpers";
 import { OlapTable } from "./olapTable";
 import { SqlResource } from "./sqlResource";
