+++ packages/ts-moose-lib/src/index.ts
@@ -1,6 +1,4 @@
-export type Key<T extends string | number | Date> = T;
-
-export type JWT<T extends object> = T;
+export * from "./browserCompatible";
 
 export type DataModelConfig<T> = Partial<{
   ingestion: true;
@@ -17,26 +15,6 @@ export * from "./blocks/helpers";
 export * from "./commons";
 export * from "./consumption-apis/helpers";
 export * from "./scripts/task";
-export {
-  Aggregated,
-  OlapTable,
-  OlapConfig,
-  Stream,
-  StreamConfig,
-  DeadLetterModel,
-  DeadLetter,
-  DeadLetterQueue,
-  IngestApi,
-  IngestConfig,
-  ConsumptionApi,
-  EgressConfig,
-  IngestPipeline,
-  SqlResource,
-  View,
-  MaterializedView,
-  Task,
-  Workflow,
-} from "./dmv2";
 
 export { createConsumptionApi } from "./consumption-apis/runner";
